pyproject.toml
[tool.poetry]
name = "brief"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
pytest = "^8.0.2"
httpx = "^0.27.0"
google-cloud-texttospeech = "^2.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
